Further changes to MyData app API
